Added tool to monitor component relationships
Fixes a major bug where any cycle within the circuit would cause the program to keep crashing indefinitely in the background. Can be expanded to check if all circuits are connected and stuff like that. I haven't incorporated it in the actual application since I'm not sure where the optimal place for it would be (probably the Application class, with a getter for the class that implements the add/remove commands?).

It's being tested and seems to be working fine, if you want to expand its functionality please don't do it yourself xoxo.